[just-doc] make function docstrings with some paragraphs more pretty
now, if a function docstring consists of some paragraphs, then each of
them will be writed in the one line, for example, the following docstring

"I love long docstrings.

So i write a big amount of the characters.  Most of them haven't
meaning.

Also, i love write docstrings with some paragraphs, this paragraphs is
the second"

will be converted into...

"I love long docstrings.

So i write a big amount of the characters.  Most of them haven't meaning.

Also, i love write docstrings with some paragraphs, this paragraphs is the second"